clip-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
transformers==4.44.0
Pillow==10.4.0
requests==2.32.0
numpy==1.26.4
python-multipart==0.0.9